application/requirements.dev.txt
-r requirements.txt
django-nose
coverage
pep8
pylint
